# docs/openapi/swagger.yaml
components:
  schemas:
...
    ? github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-github_com_Mininglamp-OSS_octo-marketplace_internal_service_ProbeResponse
    : properties:
        data:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_service.ProbeResponse'
      type: object
...
    github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-internal_api_handler_expert_SkillContentResp:
      properties:
        data:
          $ref: '#/components/schemas/internal_api_handler_expert.SkillContentResp'
      type: object
...
    github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-internal_api_handler_plugin_skillMarkdownResponse:
      properties:
        data:
          $ref: '#/components/schemas/internal_api_handler_plugin.skillMarkdownResponse'
      type: object
...
    github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-internal_api_handler_skill_SkillMDResponse:
      properties:
        data:
          $ref: '#/components/schemas/internal_api_handler_skill.SkillMDResponse'
      type: object
...
    github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error:
      properties:
        error:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.ErrorBody'
      type: object
    github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.ErrorBody:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        hint:
          type: string
        message:
          type: string
      type: object
...
    github_com_Mininglamp-OSS_octo-marketplace_internal_model.Tool:
      properties:
        description:
          type: string
        name:
          type: string
      type: object
    github_com_Mininglamp-OSS_octo-marketplace_internal_model.Transport:
      enum:
      - stdio
      - streamable-http
      - sse
      type: string
      x-enum-varnames:
      - TransportStdio
      - TransportStreamableHTTP
      - TransportSSE
...
    github_com_Mininglamp-OSS_octo-marketplace_internal_service.ProbeError:
      properties:
        code:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_service.ProbeErrorCode'
        message:
          type: string
      type: object
    github_com_Mininglamp-OSS_octo-marketplace_internal_service.ProbeErrorCode:
      enum:
      - command_not_found
      - timeout
      - init_failed
      - no_tools_capability
      type: string
      x-enum-varnames:
      - ProbeErrCommandNotFound
      - ProbeErrTimeout
      - ProbeErrInitFailed
      - ProbeErrNoToolsCapability
    github_com_Mininglamp-OSS_octo-marketplace_internal_service.ProbeRequest:
      properties:
        args:
          items:
            type: string
          type: array
          uniqueItems: false
        command:
          type: string
        env:
          additionalProperties:
            type: string
          type: object
        headers:
          additionalProperties:
            type: string
          type: object
        transport:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_model.Transport'
        url:
          type: string
      type: object
    github_com_Mininglamp-OSS_octo-marketplace_internal_service.ProbeResponse:
      properties:
        error:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_service.ProbeError'
        is_ok:
          type: boolean
        server_info:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_service.ProbeServerInfo'
        tools:
          items:
            $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_model.Tool'
          type: array
          uniqueItems: false
      type: object
    github_com_Mininglamp-OSS_octo-marketplace_internal_service.ProbeServerInfo:
      properties:
        name:
          type: string
        version:
          type: string
      type: object
...
    internal_api_handler_expert.SkillContentResp:
      properties:
        content:
          type: string
      type: object
...
    internal_api_handler_plugin.skillMarkdownResponse:
      properties:
        content:
          type: string
      type: object
...
    internal_api_handler_skill.SkillMDResponse:
      properties:
        content:
          type: string
      type: object
...
  securitySchemes:
    AdminToken:
      in: header
      name: X-Admin-Token
      type: apiKey
    Bearer:
      scheme: bearer
      type: http
info:
  contact:
    name: OCTO API Team
    url: https://github.com/Mininglamp-OSS/octo-marketplace
  description: Skill and MCP marketplace API for OCTO.
  title: Octo Marketplace API
  version: 1.0.0
openapi: 3.1.0
paths:
  /admin/mcps/_probe:
    post:
      description: Probe a remote MCP connection through the administrator surface
        without persisting it.
      operationId: admin_mcp.probe
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_service.ProbeRequest'
                description: Connection to probe
                summary: body
        description: Connection to probe
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-github_com_Mininglamp-OSS_octo-marketplace_internal_service_ProbeResponse'
          description: OK
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: VALIDATION_ERROR
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - AdminToken: []
      summary: Probe system MCP server
      tags:
      - admin_mcp
...
  /admin/skills/{skill_id}/skill_md:
    get:
      description: Return SKILL.md for a public skill without Space restriction.
      operationId: admin_skill.skillmd.get
      parameters:
      - description: Skill ID
        in: path
        name: skill_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-internal_api_handler_skill_SkillMDResponse'
          description: OK
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - AdminToken: []
      summary: Get SKILL.md (admin)
      tags:
      - admin_skill
...
  /experts/{expert_id}/skill_md:
    get:
      description: Return the stored SKILL.md text for the expert's skill at index
        i.
      operationId: expert.skill_md
      parameters:
      - description: Expert ID
        in: path
        name: expert_id
        required: true
        schema:
          type: string
      - description: Skill index (default 0)
        in: query
        name: i
        schema:
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-internal_api_handler_expert_SkillContentResp'
          description: OK
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - Bearer: []
      summary: Get expert skill content
      tags:
      - expert
...
  /plugins/skill_md:
    get:
      description: Return the SKILL.md text of a visible skill Plugin, serving both
        inlined attachments and legacy backfilled object pointers.
      operationId: plugin.skill_md
      parameters:
      - description: Plugin ID
        in: query
        name: plugin_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-internal_api_handler_plugin_skillMarkdownResponse'
          description: OK
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: VALIDATION_ERROR
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - Bearer: []
      summary: Read skill plugin SKILL.md
      tags:
      - plugin
...
  /skills/{skill_id}/skill_md:
    get:
      description: Return the SKILL.md content for the current version of a visible
        Skill.
      operationId: skill.skillmd.get
      parameters:
      - description: Skill ID
        in: path
        name: skill_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-internal_api_handler_skill_SkillMDResponse'
          description: OK
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - Bearer: []
      summary: Get SKILL.md
      tags:
      - skill
...
  /squads/{squad_id}/skill_md:
    get:
      description: Return the stored SKILL.md text for a squad member's skill at index
        i.
      operationId: squad.skill_md
      parameters:
      - description: Squad ID
        in: path
        name: squad_id
        required: true
        schema:
          type: string
      - description: Member key
        in: query
        name: member
        required: true
        schema:
          type: string
      - description: Skill index (default 0)
        in: query
        name: i
        schema:
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-internal_api_handler_expert_SkillContentResp'
          description: OK
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - Bearer: []
      summary: Get squad member skill content
      tags:
      - expert
...
servers:
- url: /v1
...